Update SubscriptionService and TokenAssignerService
