+++ server/SubscriptionService.js
@@ -16,7 +16,6 @@ const joi = require('joi');
 const express = require('express');
 const app = express();
 app.use(express.json());
-var schedule = require('node-schedule');
 
 
 //Database
@@ -37,26 +36,6 @@ const subscriptions = SubscriptionRepositoryInstance.findAll();
 
 const RateRepositoryInstance = new RateRepository()
 const rates = RateRepositoryInstance.findAll();
-let web3;
-
-// Call Job to assign tokens for the new subscribers daily on 18:00
-var j = schedule.scheduleJob('18 * * *', function () {
-  subscriptions.forEach(subscription => {
-    assignOpenFundToken(subscription);
-  });
-});
-
-/**
- * set web3 provider 
- *
- * @param {Subscription}   subscription
- * @param {web3}   web3 instance
- * 
- */
-
-function setWeb3Provider(value) {
-  web3 = value;
-}
 
 /**
  * Add Subscription for OpenFundTokens
@@ -77,7 +56,9 @@ function addSubscription(req) {
     investorId: parseInt(req.body.investorId),
     token: req.body.token,
     quantity: parseInt(req.body.quantity),
-    subScriptionDate: new Date()
+    subScriptionDate: new Date(),
+    address:req.body.address,
+    hexPrivateKey:req.body.hexPrivateKey
   };
   const result = validateSubscription(subscription);
   console.log(result);
@@ -91,8 +72,6 @@ function addSubscription(req) {
     console.log(`investor:${investor}`);
     if (investor) {
       SubscriptionRepositoryInstance.save(subscription);
-      // TODO : assignOpenFundToken call should be scheduled only once a day
-      assignOpenFundToken(subscription);
       return subscription;
     }
   } else {
@@ -123,7 +102,9 @@ function validateSubscription(subscription) {
     investorId: joi.number().required(),
     token: joi.string().min(3).required(),
     quantity: joi.number().greater(0).required(),
-    subScriptionDate: joi.allow()
+    subScriptionDate: joi.allow(),
+    address:joi.string().min(20).required(),
+    hexPrivateKey:joi.string().min(20).required()
   };
   const result = joi.validate(subscription, schema);
   if (result.error) {
@@ -156,159 +137,7 @@ function addInvestorCandidate(investor) {
   return investor;
 }
 
-/**
- * Returns the value of the given token widget.
- *
- * call NAV to get the token value in Euro
- * 
- * @param {String}   token Name
- * 
- * @return
- *   The value of the token as uint
- */
-function getNavValuationsInEuro(token) {
-  // Mock data
-  const euroValue = 100;
-  return euroValue;
-}
 
-/**
- * Returns the Amount of ethers.
- *
- * Amount of ether for the given euroValue
- * 
- * @param {integer}   assetValue        .
- * @return
- *   ether amount
- */
-function etherAmount(euroValue) {
-  // Mock data
-  var amount = 10;
-  return amount;
-}
-
-
-/**
- * Returns the transaction reciept.
- *
- * Calls OpenFundToken Smart contract and Transfer Ether to buy tokens for 
- * all subscribed investors
- * 
- * @param {Subscription}   subscription        .
- * @return
- *   transactionReciept
- */
-function assignOpenFundToken(subscription) {
-  // declare contract 
-  const OpenFund_json = require('../contracts_api/OpenFundToken.json');
-  abi = OpenFund_json.abi;
-  console.log(abi);
-  //TODO: just for test , fix address at local chain ( ganache)
-  console.log("***Instantiate OpenFundToken****");
-  var contractInstance = new web3.eth.Contract(abi, '0xef03118e3e60d9003b6c622a2e94c39ebb6985f2', {
-    from: '0x0f21f6fb13310ac0e17205840a91da93119efbec', // account 0
-    gasPrice: '20000000000' // default gas price in wei, 20 gwei in this case
-  });
-
-  var accounts;
-  var accountFrom;
-  var accountTo;
-
-  if (subscription) {
-    console.log(subscription);
-    // token value in Euro  
-    const euroValue = getNavValuationsInEuro(subscription.token);
-    console.log(euroValue);
-    //convert euroValue to ether
-    const valueInEther = etherAmount(euroValue);
-    //const valueInWei = web3.toWei(valueInEther, 'ether');
-    console.log(valueInEther);
-    // Get the accounts 
-    web3.eth.getAccounts(function (err, accs) {
-      if (err != null) {
-        console.log("There was an error fetching accounts.");
-        return;
-      }
-
-      if (accs.length == 0) {
-        console.log("Couldn't get any accounts! Make sure  Ethereum client is configured correctly.");
-        return;
-      }
-
-      accounts = accs;
-      accountFrom = accounts[0];
-      console.log(`accountFrom: ${accountFrom}`);
-      accountTo = accounts[1];
-      console.log(`accountTo: ${accountTo}`);
-
-      console.log(`contractInstance: ${contractInstance}`);
-      console.log(contractInstance.options.address);
-
-
-      // Check Token balance of accountFrom Before
-      web3.eth.call({
-        to: contractInstance.options.address,
-        data: contractInstance.methods.balanceOf(accountFrom).encodeABI()
-      }).then(balance => {
-        console.log('balance of accountFrom: ', balance);
-
-      }).catch((err) => {
-        console.log(err);
-      });
-
-      // Check Token balance of accountTo Before
-      web3.eth.call({
-        to: contractInstance.options.address,
-        data: contractInstance.methods.balanceOf(accountTo).encodeABI()
-      }).then(balance => {
-        console.log('balance of accountTo Before: ', balance);
-
-      }).catch((err) => {
-        console.log(err);
-      });
-    
-      // dry Run, Transfer tokens to accountTo 
-      web3.eth.call({
-        to: contractInstance.options.address,
-        data: contractInstance.methods.transfer(accountTo, 1 ).encodeABI()
-      }).then(balance => {
-        console.log('balance of accountTo: ', balance);
-
-      }).catch((err) => {
-        console.log(err);
-      });
-
-
-       // sendTransaction, Transfer tokens to accountTo 
-
-      web3.eth.sendTransaction({
-        to: contractInstance.options.address,
-        data: contractInstance.methods.transfer(accountTo, 1 ).encodeABI()
-      }).then(reciept => {
-        console.log('TRX reciept: ', reciept);
-
-      }).catch((err) => {
-        console.log(err);
-      });
-
-       // Check Token balance of accountTo After
-       web3.eth.call({
-        to: contractInstance.options.address,
-        data: contractInstance.methods.balanceOf(accountTo).encodeABI()
-      }).then(balance => {
-        console.log('balance of accountTo After: ', balance);
-
-      }).catch((err) => {
-        console.log(err);
-      });
-
-    });
-
-
-
-
-  }
-}
 
 /**
  * Returns all subscriptions.
@@ -348,7 +177,6 @@ function getFunds() {
 
 
 module.exports = {
-  setWeb3Provider: setWeb3Provider,
   validateSubscription: validateSubscription,
   addSubscription: addSubscription,
   getSubscriptions: getSubscriptions,
